vectorize copy of delta length byte arrays
