+++ encoding/delta/length_byte_array_amd64.go
@@ -5,6 +5,7 @@ package delta
 import (
 	"fmt"
 
+	"github.com/segmentio/parquet-go/encoding/plain"
 	"golang.org/x/sys/cpu"
 )
 
@@ -42,4 +43,41 @@ func decodeLengthValues(lengths []int32) (int, error) {
 		sum, err = decodeLengthValuesDefault(lengths)
 	}
 	return sum, err.check()
+}
+
+//go:noescape
+func decodeLengthByteArrayAVX2(dst, src []byte, lengths []int32)
+
+func decodeLengthByteArray(dst, src []byte, lengths []int32, totalLength int) {
+	i := 0
+	j := 0
+	k := 0
+	n := 0
+
+	// To leverage the SEE optimized implementation of the function we must
+	// create enough padding at the end to prevent the opportunisitic reads
+	// and writes from overflowing past the buffers limits.
+	if cpu.X86.HasAVX2 && totalLength > lengthByteArrayPadding {
+		k = len(lengths)
+
+		for k > 0 && n < lengthByteArrayPadding {
+			k--
+			n += int(lengths[k])
+		}
+
+		if k > 0 && n >= lengthByteArrayPadding {
+			decodeLengthByteArrayAVX2(dst, src, lengths[:k])
+			j = totalLength - n
+			i = plain.ByteArrayLengthSize*k + j
+		} else {
+			k = 0
+		}
+	}
+
+	for _, n := range lengths[k:] {
+		plain.PutByteArrayLength(dst[i:], int(n))
+		i += plain.ByteArrayLengthSize
+		i += copy(dst[i:], src[j:j+int(n)])
+		j += int(n)
+	}
 }
