+++ encoding/delta/length_byte_array_purego.go
@@ -2,6 +2,8 @@
 
 package delta
 
+import "github.com/segmentio/parquet-go/encoding/plain"
+
 func decodeLengthValues(lengths []int32) (sum int, err error) {
 	for _, n := range lengths {
 		sum += int(n)
@@ -10,4 +12,16 @@ func decodeLengthValues(lengths []int32) (sum int, err error) {
 		}
 	}
 	return sum, nil
+}
+
+func decodeLengthByteArray(dst, src []byte, lengths []int32, _ int) {
+	i := 0
+	j := 0
+
+	for _, n := range lengths {
+		plain.PutByteArrayLength(dst[i:], int(n))
+		i += plain.ByteArrayLengthSize
+		i += copy(dst[i:], src[j:j+int(n)])
+		j += int(n)
+	}
 }
